refactor: move autocomplete func to utils folder
